add: serializer function and move errors to directory
