+++ lib/errors/index.js
@@ -1,5 +1,6 @@
 /*jshint esversion:6, node:true*/
 'use strict';
+const addErrorSerializer = require('./serializer');
 
 module.exports.init = function(app, config) {
     //TODO: Integrate signals here!
@@ -9,26 +10,7 @@ module.exports.init = function(app, config) {
      * using our regular JSON class.
      */
     if (!('toJSON' in Error.prototype)) {
-        Object.defineProperty(Error.prototype, 'toJSON', {
-            value: function() {
-                var alt = {};
-
-                Object.getOwnPropertyNames(this).forEach(function(key) {
-                    alt[key] = this[key];
-                }, this);
-
-                /**
-                 * Remove stack trace from production environments
-                 */
-                if (app.isProduction) {
-                    delete alt.stack;
-                }
-
-                return alt;
-            },
-            writable: true,
-            configurable: true,
-        });
+        addErrorSerializer(app.isProduction);
     }
 
     /*
